+++ main.py
@@ -103,7 +103,12 @@ def afterTurn(player):
                                                "street":fc.f_container[player.fid]} )
                 # st_transact.buyStreet(player, fc.f_container[player.fid])
             elif fc.f_container[player.fid].owner == player:
-                break # Insert code to buy stuff here
+                trade_phase = offer.house_hotel_card(screen, 
+                                               phase = trade_phase,
+                                               ftc_house= st_transact.buyHouse, 
+                                               ftc_hotel= st_transact.buyHotel,
+                                               kw_args={"player":player,
+                                               "street":fc.f_container[player.fid]} )
             
             else:
                 st_transact.payRent(player, fc.f_container[player.fid])