+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This build is not intended for submission, but rather for the developers to keep
 - "Misonic Project" is now called "Dynopoly".
 - In the side bar, the label "Misonic Project" now shows "Current Player Info" instead.
 - The window title is now "Dynopoly".
+- You can now buy a house and a hotel.
     
 ## Known issues